Modularize app to support testing
