+++ server/db/index.js
@@ -4,7 +4,7 @@ require('dotenv').config()
 
 const pool = new Pool({
   connectionString:
-    process.env.NODE_ENV === 'testing'
+    process.env.NODE_ENV === 'test'
       ? process.env.TEST_DATABASE_URL
       : process.env.DATABASE_URL,
   ssl:
@@ -14,6 +14,7 @@ const pool = new Pool({
         }
       : undefined,
 })
+
 const query = (text, params) => pool.query(text, params)
 /**
  * @returns {Promise<pg.PoolClient>}
